address shard singletons
